backend/requirements.txt
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
python-dotenv==1.2.2